requirements.txt
numpy<2.0.0  # Pin numpy to 1.x version to avoid compatibility issues
torch>=2.0.0
transformers>=4.30.0
google-cloud-storage>=2.10.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
huggingface_hub==0.19.4
accelerate==0.24.1 